EcoDetectWeb/requirements.txt
streamlit
pymongo
python-dotenv
fastapi
uvicorn[standard]